change prefix from 'encode_' to 'enc_'
using the 'encode_' prefix will lead to names like:

  encode_utf8_encode()

which sounds and looks confusing and overly verbose.
The same thing happens with decoding functions:

  encode_utf8_decode()

a newcomer to libencode might wonder what this function actually does.
If names aren't intuitive and fail to clearly express meaning, they are bad.
A prefix especially shouldn't lead to that much confusion since it's only
purpose is to delimit the namespace of the library.