utf8: always return the amount of bytes needed for an encoding in utf8_encode()
Providing the user with the amount of bytes that *would* have been needed
to properly encode (and store) the given codepoint is way more useful than
just returning zero. The user is now able to calculate the missing bytes
more easily and can also use utf8_decode() as a length function which
improves the interface (and functionality) a bit.